Pull Configs from File
These files are used to pull configurations from the attached file to apply the configuration to all switches. This scenario deviates from David's example by using one file per device rather the one file for the core switches and one file for the leaf switches. There is a python script to add the file config, and another to delete the actions taken.